requirements.txt
astronomer-cosmos==1.10.0
apache-airflow-providers-snowflake
dbt-snowflake==1.11.2
